Move _get_ui_registry from ui_tester.py to inside get_default_registry
